Really fix rendering issue with "curve drag point".

+++ src/ui/tool/path-manipulator.cpp
@@ -1696,7 +1696,7 @@ Geom::Coord PathManipulator::_updateDragPoint(Geom::Point const &evp)
     {
         _dragpoint->setVisible(true);
         _dragpoint->setPosition(_desktop->w2d(nearest_pt));
-        _dragpoint->setSize(2 * stroke_tolerance - 1);
+        _dragpoint->setSize(2 * (int)stroke_tolerance - 1); // stroke_tolerance is at least two.
         _dragpoint->setTimeValue(fracpart);
         _dragpoint->setIterator(first);
     } else {
